Release version 1.1.0
Fix a bug when switching modes. This error occurred in two scenarios:
Scenario 1
1- Being in integrated mode.
2- Switch to hybrid, and not reboot.
3- Try to switch to nvidia without success (an EnvyControl error occurs).

Scenario 2
1- Being in integrated mode.
2- Try to switch to nvidia without success (an EnvyControl error occurs)

In both scenarios EnvyControl automatically switches to hybrid mode without any clear or explicit warning.
This left the widget out of sync with what was actually going on.
Note that when switching from integrated to nvidia it is necessary to first switch to hybrid, reboot, and then switch to nvidia.

Add a button to refresh the state of the widget.

Minor bug fixes and improvements.